Hunterize J.D. Koftinoff Software, Ltd's C++ MIDI Library
- Overhaul CMake files to allow for easier importing of library into other CMake based projects.
- Add version information and pre-processed version header file to expose the library version in code.
- Change building of utility and test applications to be optionally disabled by default.
- Add custom CMake source file source groups.
- Remove unnecessary pre-generated IDE project files.
- Add gitignore file to prevent undesireable files from being committed.
- Fix string length data loss build warning due to lack of explicit cast operation.

+++ src/jdksmidi_filewrite.cpp
@@ -325,7 +325,7 @@ void MIDIFileWrite::WriteTextEvent ( unsigned long time, unsigned char type, con
     WriteCharacter ( type ); // Text event type
     IncrementCounters ( 2 );
 
-    int len = strlen ( text );
+    int len = static_cast<int>( strlen ( text ) );
     IncrementCounters ( WriteVariableNum ( len ) );
 
     while ( *text )
